backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.19
pdfplumber==0.11.4
python-docx==1.1.2
google-generativeai==0.8.3
supabase==2.11.0
pgvector==0.3.6
httpx==0.28.1
python-dotenv==1.0.1
unstructured-client==0.26.0
